+++ example/plugins/bootstrap.js
@@ -1,3 +1,0 @@
-if (process.BROWSER_BUILD) {
-  require('bootstrap');
-}