+++ example/plugins/vendor.js
@@ -1,0 +1,4 @@
+if (process.browser) {
+  require('expose-loader?Choices!choices.js');
+  require('choicesjs-stencil/dist/choicesjsstencil');
+}